+++ lib/widgets/collection/app_bar.dart
@@ -254,6 +254,7 @@ class _CollectionAppBarState extends State<CollectionAppBar> with SingleTickerPr
                       _buildRotateAndFlipMenuItems(context, canApply: canApply),
                       ...[
                         EntrySetAction.editDate,
+                        EntrySetAction.editLocation,
                         EntrySetAction.editRating,
                         EntrySetAction.editTags,
                         EntrySetAction.removeMetadata,
@@ -439,6 +440,7 @@ class _CollectionAppBarState extends State<CollectionAppBar> with SingleTickerPr
       case EntrySetAction.rotateCW:
       case EntrySetAction.flip:
       case EntrySetAction.editDate:
+      case EntrySetAction.editLocation:
       case EntrySetAction.editRating:
       case EntrySetAction.editTags:
       case EntrySetAction.removeMetadata: